+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-## [v2.14] - 2025-12-03
+## [v2.15] - 2025-12-03
 
 ## Added 
 - `SeatPurchasingFeeInstruction` - for fee alignment